refactor: Convert academics-shared/ to TypeScript
Converted all 18 shared course files from JSON to TypeScript for
complete format consistency across the entire codebase.

Changes:
- Converted 18 files: .json → .ts in academics-shared/
- Updated 11 section files: removed .json from import paths
- Updated Pages.tsx: ResearchOverview import
- Deleted old JSON files

Files converted:
- All course definitions (COSC, ECE, MATH, PSYC)
- ResearchOverview
- All now TypeScript with proper exports

Benefits:
- ✅ 100% TypeScript codebase (no JSON/TS mixing)
- ✅ Consistent architecture throughout
- ✅ Type safety for all course definitions
- ✅ Cleaner imports (no .json extensions)
- ✅ Future-proof for enhancements

Impact:
- academics-shared/: All TypeScript ✓
- academics/: All TypeScript ✓
- academics-pyramid/: All TypeScript ✓
- academics-domains/: All TypeScript ✓
- config/: Updated imports ✓

Validated: TypeScript compilation passes with no errors.
