+++ src/data/academics-domains/SystemsArchitecture.ts
@@ -5,10 +5,10 @@ Github Repository: https://github.com/andreicozma1/andreicozma1.github.io
 */
 
 // Import shared course definitions
-import cosc530 from '../academics-shared/COSC530-ComputerSystems.json'
-import cosc562 from '../academics-shared/COSC562-OperatingSystems.json'
-import cosc581 from '../academics-shared/COSC581-Algorithms.json'
-import ece511 from '../academics-shared/ECE511-LinearSystems.json'
+import cosc530 from '../academics-shared/COSC530-ComputerSystems'
+import cosc562 from '../academics-shared/COSC562-OperatingSystems'
+import cosc581 from '../academics-shared/COSC581-Algorithms'
+import ece511 from '../academics-shared/ECE511-LinearSystems'
 
 export default {
 	title: "Systems Architecture & Design",
