+++ src/data/academics-pyramid/GraduateML.ts
@@ -5,11 +5,11 @@ Github Repository: https://github.com/andreicozma1/andreicozma1.github.io
 */
 
 // Import shared course definitions
-import cosc522 from '../academics-shared/COSC522-MachineLearning.json'
-import cosc523 from '../academics-shared/COSC523-ArtificialIntelligence.json'
-import cosc525 from '../academics-shared/COSC525-DeepLearning.json'
-import ece517 from '../academics-shared/ECE517-ReinforcementLearning.json'
-import cosc524 from '../academics-shared/COSC524-NLP.json'
+import cosc522 from '../academics-shared/COSC522-MachineLearning'
+import cosc523 from '../academics-shared/COSC523-ArtificialIntelligence'
+import cosc525 from '../academics-shared/COSC525-DeepLearning'
+import ece517 from '../academics-shared/ECE517-ReinforcementLearning'
+import cosc524 from '../academics-shared/COSC524-NLP'
 
 export default {
 	title: "Tier 2: Intelligent Systems & Machine Learning",
