+++ src/data/academics-shared/COSC559-HCI.ts
@@ -1,4 +1,10 @@
-{
+/*
+Author: Andrei Cozma
+Website: https://andreicozma.com
+Github Repository: https://github.com/andreicozma1/andreicozma1.github.io
+*/
+
+export default {
 	"title": "COSC 559 - Human-Computer Interaction",
 	"subtitle": "Fundamentals of HCI examining how human factors affect interface design and usability, covering perception, cognition, and user-centered design methodologies.",
 	"chips": {
