+++ src/data/academics-shared/COSC581-Algorithms.ts
@@ -1,4 +1,10 @@
-{
+/*
+Author: Andrei Cozma
+Website: https://andreicozma.com
+Github Repository: https://github.com/andreicozma1/andreicozma1.github.io
+*/
+
+export default {
 	"title": "COSC 581 - Algorithms",
 	"subtitle": "Advanced design and analysis of algorithms with focus on asymptotics, relative efficiencies, and implementation strategies. Covers algorithm design techniques, computational complexity, and selected advanced topics.",
 	"chips": {
