+++ src/data/academics-shared/ECE511-LinearSystems.ts
@@ -1,4 +1,10 @@
-{
+/*
+Author: Andrei Cozma
+Website: https://andreicozma.com
+Github Repository: https://github.com/andreicozma1/andreicozma1.github.io
+*/
+
+export default {
 	"title": "ECE 511 - Linear Systems Theory",
 	"subtitle": "State space models of linear dynamical systems, covering linear algebra, state transition, controllability, observability, realization theory, and stability analysis.",
 	"chips": {
