+++ src/data/academics-shared/ECE517-ReinforcementLearning.ts
@@ -1,4 +1,10 @@
-{
+/*
+Author: Andrei Cozma
+Website: https://andreicozma.com
+Github Repository: https://github.com/andreicozma1/andreicozma1.github.io
+*/
+
+export default {
 	"title": "ECE 517 - Reinforcement Learning",
 	"subtitle": "Principles and methods for reinforcement learning and sequential stochastic control; Markov decision problems; dynamic programming; temporal difference learning; design considerations for hardware and software.",
 	"chips": {
