+++ src/data/academics-shared/ResearchOverview.ts
@@ -1,4 +1,10 @@
-{
+/*
+Author: Andrei Cozma
+Website: https://andreicozma.com
+Github Repository: https://github.com/andreicozma1/andreicozma1.github.io
+*/
+
+export default {
 	"title": "Research & Academic Status",
 	"variant": "timeline",
 	"items": [
