+++ src/data/academics/IntelligentSystems.ts
@@ -5,10 +5,10 @@ Github Repository: https://github.com/andreicozma1/andreicozma1.github.io
 */
 
 // Import shared course definitions
-import cosc525 from '../academics-shared/COSC525-DeepLearning.json'
-import cosc523 from '../academics-shared/COSC523-ArtificialIntelligence.json'
-import cosc522 from '../academics-shared/COSC522-MachineLearning.json'
-import cosc420 from '../academics-shared/COSC420-BiologicallyInspired.json'
+import cosc525 from '../academics-shared/COSC525-DeepLearning'
+import cosc523 from '../academics-shared/COSC523-ArtificialIntelligence'
+import cosc522 from '../academics-shared/COSC522-MachineLearning'
+import cosc420 from '../academics-shared/COSC420-BiologicallyInspired'
 
 export default {
 	title: "Intelligent Systems",
